refactor: decouple charts from ember
before this change to build a chart you had to create a boat load of
files because rendering was done in ember world but configuring was done
in js world.

This change brings world and configure into js. as long as a user
creates a render function it will work.

TODO:
decide how to handle table
make the default the render function
