+++ app/components/chart-editor.hbs
@@ -24,7 +24,7 @@
             @error={{@chart.dataset.error}}
           as |size| >
             {{#if (and @chart.dataset.cache.results @chart.dataset.cache.results.length)}}
-                {{component @chart.visualisation.component options=@chart.options  results=@chart.dataset.cache.results width=size.width height=size.height}}
+                {{component @chart.visualisation.component options=@chart.options  results=@chart.dataset.cache.results width=size.width height=size.height visualisation=@chart.visualisation}}
             {{/if}}
           </Widget>
         </Dashboard>
