--- app/components/viz/area.hbs
@@ -1,1 +1,0 @@
-<Vega @schema={{this.schema}}/>
