--- app/components/viz/bar.hbs
@@ -1,1 +1,0 @@
-<Vega @schema={{this.schema}}/>
