+++ app/components/viz/control-panel.hbs
