+++ app/components/viz/control-panel.js
