--- app/components/viz/line.hbs
@@ -1,1 +1,0 @@
-<Vega @schema={{this.schema}}/>
