+++ app/components/viz/render-func.hbs
@@ -1,0 +1,1 @@
+  <div {{update-on-data-change @results @visualisation @options @height}}></div>
