+++ app/modifiers/update-on-data-change.js
@@ -1,0 +1,14 @@
+import { modifier } from 'ember-modifier';
+
+export default modifier(function updateOnSchema(
+  element,
+  [results, visualisationModel, options, height]
+) {
+  try {
+    visualisationModel
+      .get('raw')
+      .render(element, results, { ...options, height });
+  } catch {
+    element.innerHTML = 'something went wrong';
+  }
+});
